Added full temperature and humidity readings.

+++ src/data_bus.c
@@ -2,5 +2,5 @@
 
 uint16 V_Temperature = 0;
 
-uint16 V_Humidity = 0;
+uint8  V_Humidity = 0;
