+++ src/data_bus.h
@@ -4,7 +4,7 @@
 #include "pico_types.h"
 
 extern uint16 V_Temperature;
-extern uint16 V_Humidity;
+extern uint8  V_Humidity;
 
 static inline uint16 GetTemperature(void) {
     return V_Temperature;
@@ -14,11 +14,11 @@ static inline void SetTemperature(uint16 Temperature) {
     V_Temperature = Temperature;
 }
 
-static inline uint16 GetHumidity(void) {
+static inline uint8 GetHumidity(void) {
     return V_Humidity;
 }
 
-static inline void SetHumidity(uint16 Humidity) {
+static inline void SetHumidity(uint8 Humidity) {
     V_Humidity = Humidity;
 }
